Update and rename Group_42_R_Code.Rmd to Customer transaction prediction.Rmd
Considering this is highly imbalanced data, accuracy is not a good evaluation metric to check the model performance (even for the initial analysis) for this specific data mining problem (minority class is much more important than the majority class). The bank would like to identify true positives correctly (customers who will make the specific transaction). Therefore, even in the initial evaluation, it is better not to focus on accuracy. Recall is an important measure that needs to be checked. Precision also provides information on the false positives. These two evaluation techniques should be considered. It would be better if we could interpret the numerical results considering the objective of the bank.